pyproject.toml
[tool.poetry]
name = "frcm"
version = "0.1.0"
description = ""
authors = ["Lars Michael Kristensen <lmkr@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "frcm", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.5.3"
python-dateutil = "^2.8.2"
pytest = "^7.4.4"
numpy = "^1.26.3"
requests = "^2.31.0"
python-decouple = "^3.8"
fastapi = {extras = ["all"], version = "^0.109.2"}
pyopenssl = "^24.0.0"
psycopg2-binary = "^2.9.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
